+++ src/core/game/index.ts
@@ -1,0 +1,6 @@
+
+import "./teamSelection";
+import "./beat";
+import "./start";
+import "./play";
+import "./end";