+++ src/index.ts
@@ -1,9 +1,10 @@
 
-import "core/game";
+import "core/game/index";
 import "core/init";
 import "egg/dolly";
 import "egg/gem";
 import "egg/sparkles";
+import "misc/abilityPreload";
 import "misc/angle";
 import "misc/clean";
 import "misc/control";