+++ src/shared.ts
@@ -2,7 +2,6 @@
 // This file should not import anything from this repo except libs
 
 import { addScriptHook, W3TS_HOOK } from "@voces/w3ts";
-import { AbilityRangePreload } from "./misc/abilityPreload";
 
 export const getterSetterFunc = <T>( init?: T ): ( newValue?: T ) => T => {
 
@@ -20,8 +19,6 @@ export const getterSetterFunc = <T>( init?: T ): ( newValue?: T ) => T => {
 	};
 
 };
-
-type GAME_STATES = "init" | "start" | "play";
 
 export const fillArray = <T>( size: number, value: T, arr: Array<T> = [] ): Array<T> => {
 
@@ -41,13 +38,74 @@ export const fillArrayFn = <T>( size: number, fn: ( index: number ) => T, arr: A
 
 };
 
-export const color = [
-	"|CFFFF0303", "|CFF0042FF", "|CFF1CE6B9", "|CFF540081",
-	"|CFFFFFF01", "|CFFFE8A0E", "|CFF20C000", "|CFFE55BB0",
-	"|CFF959697", "|CFF7EBFF1", "|CFF106246", "|CFF4E2A04",
-	"|CFF3F81F8", "|CFFC00040", "|CFFD9D919",
-];
-export const gameState: ( newState?: GAME_STATES ) => GAME_STATES = getterSetterFunc( "init" as GAME_STATES );
+export const color = {
+	red: "|cffff0303",
+	blue: "|cff0042ff",
+	teal: "|cff1ce6b9",
+	purple: "|cff540081",
+	yellow: "|cfffffc00",
+	orange: "|cfffe8a0e",
+	green: "|cff20c000",
+	pink: "|cffe55bb0",
+	gray: "|cff959697",
+	lightblue: "|cff7ebff1",
+	darkgreen: "|cff106246",
+	brown: "|cff4a2a04",
+
+	maroon: "|cff9b0000",
+	navy: "|cff0000c3",
+	turquoise: "|cff00eaff",
+	violet: "|cffbe00fe",
+	wheat: "|cffebcd87",
+	peach: "|cfff8a48b",
+	mint: "|cffbfff80",
+	lavender: "|cffdcb9eb",
+	coal: "|cff282828",
+	snow: "|cffebf0ff",
+	emerald: "|cff00781e",
+	peanut: "|cffa46f33",
+
+	sheepblue: "|CFF3F81F8",
+	wolfred: "|CFFC00040",
+	gold: "|CFFD9D919",
+};
+
+export type Color = keyof typeof color;
+
+const wc3ColorMap = new Map();
+wc3ColorMap.set( PLAYER_COLOR_RED, "red" );
+wc3ColorMap.set( PLAYER_COLOR_BLUE, "blue" );
+wc3ColorMap.set( PLAYER_COLOR_CYAN, "teal" );
+wc3ColorMap.set( PLAYER_COLOR_PURPLE, "purple" );
+wc3ColorMap.set( PLAYER_COLOR_YELLOW, "yellow" );
+wc3ColorMap.set( PLAYER_COLOR_ORANGE, "orange" );
+wc3ColorMap.set( PLAYER_COLOR_GREEN, "green" );
+wc3ColorMap.set( PLAYER_COLOR_PINK, "pink" );
+wc3ColorMap.set( PLAYER_COLOR_LIGHT_GRAY, "gray" );
+wc3ColorMap.set( PLAYER_COLOR_LIGHT_BLUE, "lightblue" );
+wc3ColorMap.set( PLAYER_COLOR_AQUA, "darkgreen" );
+wc3ColorMap.set( PLAYER_COLOR_BROWN, "brown" );
+wc3ColorMap.set( PLAYER_COLOR_MAROON, "maroon" );
+wc3ColorMap.set( PLAYER_COLOR_NAVY, "navy" );
+wc3ColorMap.set( PLAYER_COLOR_TURQUOISE, "turquoise" );
+wc3ColorMap.set( PLAYER_COLOR_VIOLET, "violet" );
+wc3ColorMap.set( PLAYER_COLOR_WHEAT, "wheat" );
+wc3ColorMap.set( PLAYER_COLOR_PEACH, "peach" );
+wc3ColorMap.set( PLAYER_COLOR_MINT, "mint" );
+wc3ColorMap.set( PLAYER_COLOR_LAVENDER, "lavender" );
+wc3ColorMap.set( PLAYER_COLOR_COAL, "coal" );
+wc3ColorMap.set( PLAYER_COLOR_SNOW, "snow" );
+wc3ColorMap.set( PLAYER_COLOR_EMERALD, "emerald" );
+wc3ColorMap.set( PLAYER_COLOR_PEANUT, "peanut" );
+
+export const playerColorToColor = ( playerColor: playercolor ): Color => {
+
+	const color = wc3ColorMap.get( playerColor );
+	if ( ! color ) throw `unknown color ${playerColor}`;
+	return color;
+
+};
+
 export const goldFactor: ( newFactory?: number ) => number = getterSetterFunc( 1 );
 export const isHere = (): boolean => GetPlayerSlotState( GetFilterPlayer() ) === PLAYER_SLOT_STATE_PLAYING;
 export const saveskills: Array<number> = fillArray( bj_MAX_PLAYERS, 0 );
@@ -73,7 +131,10 @@ let someInteger: number;
 
 const countHereEnum = (): void => {
 
-	if ( GetPlayerSlotState( GetEnumPlayer() ) === PLAYER_SLOT_STATE_PLAYING )
+	if (
+		GetPlayerSlotState( GetEnumPlayer() ) === PLAYER_SLOT_STATE_PLAYING ||
+		GetPlayerSlotState( GetEnumPlayer() ) === PLAYER_SLOT_STATE_EMPTY
+	)
 		someInteger = someInteger + 1;
 
 };
@@ -181,7 +242,7 @@ export const wolfUnit = ( p: player ): unit => {
 
 };
 
-export const SmallText = ( amount: number, u: unit, cc: number, x: number, y: number ): void => {
+export const SmallText = ( amount: number, u: unit, cc: Color, x: number, y: number ): void => {
 
 	if ( GetUnitAbilityLevel( u, FourCC( "Alv1" ) ) <= 0 && IsVisibleToPlayer( GetUnitX( u ), GetUnitY( u ), GetLocalPlayer() ) ) {
 
@@ -202,9 +263,6 @@ addScriptHook( W3TS_HOOK.MAIN_BEFORE, (): void => {
 
 	SetMapFlag( MAP_SHARED_ADVANCED_CONTROL, true );
 
-	ForceEnumAllies( sheepTeam, Player( 0 ), Condition( isHere ) );
-	ForceEnumAllies( wolfTeam, Player( 11 ), Condition( isHere ) );
-
 } );
 
 export const BLACK_SHEEP_TYPE = FourCC( "uC02" );
@@ -219,9 +277,3 @@ export const isUnitSheep = ( unit: unit ): boolean =>
 
 export const isUnitWolf = ( unit: unit ): boolean =>
 	wolfTypes.includes( GetUnitTypeId( unit ) );
-
-addScriptHook( W3TS_HOOK.MAIN_AFTER, (): void => {
-
-	AbilityRangePreload( FourCC( "A001" ), FourCC( "A00Q" ) );
-
-} );