+++ src/util/log.ts
@@ -27,6 +27,8 @@ export const termToString = ( v: any, color = true ): string => {
 	if ( typeof v === "number" ) return color ? colorize.number( v ) : v.toString();
 	if ( typeof v === "boolean" ) return color ? colorize.boolean( v ) : v.toString();
 	if ( typeof v === "function" ) return color ? colorize.number( "[function]" ) : "[function]";
+	if ( v === undefined ) return color ? colorize.boolean( "undefined" ) : "undefined";
+	if ( v == null ) return color ? colorize.boolean( "null" ) : "null";
 
 	if ( isArray( v ) ) {
 
@@ -41,15 +43,19 @@ export const termToString = ( v: any, color = true ): string => {
 	if ( typeof v === "object" && v != null )
 		return `{ ${Object.entries( v ).map( ( [ key, value ] ) => `${key}: ${termToString( value )}` ).join( ", " )} }`;
 
-	if ( v === undefined ) return color ? colorize.boolean( "undefined" ) : "undefined";
-	if ( v == null ) return color ? colorize.boolean( "null" ) : "null";
-
 	const type = userdataType( v );
 
 	switch ( type ) {
 
 		case "player": return `Player ${termToString( { id: GetPlayerId( v ), name: GetPlayerName( v ) } )}`;
 		case "unit": return `Unit ${termToString( { name: GetUnitName( v ), owner: GetOwningPlayer( v ) } )}`;
+		case "force": {
+
+			const arr: player[] = [];
+			ForForce( v, () => arr.push( GetEnumPlayer() ) );
+			return `Force ${termToString( arr )}`;
+
+		}
 		default: {
 
 			let handleId = - 1;